requirements.txt
# pip install -r requirements.txt

torch>=2.0
torch-geometric
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv

numpy>=1.24.0
pandas>=1.1.0
scipy>=1.5.0
scikit-learn>=0.24.0

matplotlib>=3.4.0
imageio>=2.9.0

tqdm>=4.60.0
pyyaml>=5.4.0

ase>=3.22.0

pytest>=6.2.4

tensorboard>=2.10.0